fix: restructure Shopware Services test
